Base Three NPC Done

+++ Source/Student/Project_1/Decorator/D_KeepTryingUntilFourTimesOrSuccess.h
@@ -7,7 +7,7 @@ public:
     D_KeepTryingUntilFourTimesOrSuccess();
 
 protected:
-    unsigned counter;
+    float counter;
 
     virtual void on_enter() override;
     virtual void on_update(float dt) override;
